+++ tabular_permissions/admin.py
@@ -8,18 +8,18 @@ from . import app_settings
 User = get_user_model()
 
 
-class UserTabularPermissionsAdminBase(object):
+class UserTabularPermissionsMixin(object):
     def formfield_for_manytomany(self, db_field, request=None, **kwargs):
-        field = super(UserTabularPermissionsAdminBase, self).formfield_for_manytomany(db_field, request, **kwargs)
+        field = super(UserTabularPermissionsMixin, self).formfield_for_manytomany(db_field, request, **kwargs)
         if db_field.name == 'user_permissions':
             field.widget = TabularPermissionsWidget(db_field.verbose_name, db_field.name in self.filter_vertical)
             field.help_text = ''
         return field
 
 
-class GroupTabularPermissionsAdminBase(object):
+class GroupTabularPermissionsMixin(object):
     def formfield_for_manytomany(self, db_field, request=None, **kwargs):
-        field = super(GroupTabularPermissionsAdminBase, self).formfield_for_manytomany(db_field, request, **kwargs)
+        field = super(GroupTabularPermissionsMixin, self).formfield_for_manytomany(db_field, request, **kwargs)
         if db_field.name == 'permissions':
             field.widget = TabularPermissionsWidget(db_field.verbose_name, db_field.name in self.filter_vertical,
                                                     'permissions')
@@ -27,15 +27,15 @@ class GroupTabularPermissionsAdminBase(object):
         return field
 
 
-class TabularPermissionsUserAdmin(UserTabularPermissionsAdminBase, UserAdmin):
+class TabularPermissionsUserAdmin(UserTabularPermissionsMixin, UserAdmin):
     pass
 
 
-class TabularPermissionsGroupAdmin(GroupTabularPermissionsAdminBase, GroupAdmin):
+class TabularPermissionsGroupAdmin(GroupTabularPermissionsMixin, GroupAdmin):
     pass
 
 
-if app_settings.TABULAR_PERMISSIONS_AUTO_IMPLEMENT:
+if app_settings.AUTO_IMPLEMENT:
     try:
         admin.site.unregister(User)
         admin.site.register(User, TabularPermissionsUserAdmin)